+++ qmt/__init__.py
@@ -1,4 +1,3 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License.
 from .physics_constants import units, constants, parse_unit, to_float
-from .materials import Materials
