+++ qmt/geometry/builder_3d.py
@@ -1,183 +1,55 @@
-from enum import Enum
-from dataclasses import dataclass, field
+"""
+The Geo3DBuilder class, which is used to build 3D geometries
+"""
+
 from typing import List, Optional
-from qmt.data.data_utils import write_deserialised
+from qmt.data import serialize_file, write_deserialised
+import FreeCAD
+from .part_3d import (
+    DepoMode,
+    ExtrudeData,
+    LithographyData,
+    Part3DData,
+    SAGData,
+    WireData,
+    WireShellData,
+)
+from qmt.geometry import Geo3DData
 
 
-class DomainType(Enum):
-    """
-    - semiconductor -- region permitted to self-consistently accumulate
-    - metal_gate -- an electrode
-    - virtual -- a part just used for selection (ignores material)
-    - dielectric -- no charge accumulation allowed
-    """
-
-    semiconductor = 1
-    metal_gate = 2
-    virtual = 3
-    dielectric = 4
-
-
-@dataclass
-class Part3DData:
-    """
-    Base class for a 3D geometric part.
-    :param domain_type: The type of domain this part represents
-    :param fc_name: The name of the 2D/3D freeCAD object that this is built from. Note
-        that if the label used for the 3D part is the same as the freeCAD label, and that
-        label is unique, None may be used here as a shortcut
-    :param label: The descriptive name of this new part
-    :param material: The material of the resulting part.
-    """
-
-    built_fc_name: Optional[str] = field(init=False)  # This gets set on geometry build
-    domain_type: DomainType
-    fc_name: str
-    label: str
-    material: Optional[str]
-    serial_stp: Optional[str] = field(init=False)  # This gets set on geometry build
-
-    def __post_init__(self):
-        if not self.material and self.domain_type == DomainType.virtual:
-            raise ValueError(
-                f"Material must be provided for non virtual ({self.domain_type.name}) "
-                f"part {self.label}"
-            )
-        elif self.material and self.domain_type == DomainType.virtual:
-            raise ValueError(f"Material provided for virtual part {self.label}")
-
-    def write_stp(self, file_path=None):
-        """Write part geometry to a STEP file.
-
-        Returns the STEP file path.
-        """
-        if file_path is None:
-            file_path = f"{self.label}.stp"
-        write_deserialised(self.serial_stp, file_path)
-        return file_path
-
-
-@dataclass
-class ExtrudeData(Part3DData):
-    """
-    :param thickness: The extrusion thickness
-    :param z0: The starting z coordinate
-    """
-
-    thickness: float
-    z0: float
-
-
-@dataclass
-class WireData(Part3DData):
-    """
-    :param thickness: The total wire thickness
-    :param z0: The starting z coordinate
-    """
-
-    thickness: float
-    z0: float
-
-
-class DepoMode(Enum):
-    """
-    'depo' or 'etch' defines the positive or negative mask for the deposition of a wire
-    coating
-    """
-
-    depo = 1
-    etch = 2
-
-
-@dataclass
-class WireShellData(Part3DData):
-    """
-    :param thickness: The wire shell thickness
-    :param target_wire: Target wire for coating
-    :param shell_verts: Vertices to use when rendering the coating
-    :param depo_mode: 'depo' or 'etch' defines the positive or negative mask
-    """
-
-    thickness: float
-    target_wire: WireData
-    shell_verts: List[int]
-    depo_mode: DepoMode
-
-
-@dataclass
-class SAGData(Part3DData):
-    """
-    :param thickness: The total SAG thickness
-    :param z0: The starting z coordinate
-    :param z_middle: The location for the "flare out"
-    :param t_in: The lateral distance from the 2D profile to the edge of the top bevel
-    :param t_out: The lateral distance from the 2D profile to the furthest "flare out"
-        location
-    """
-
-    thickness: float
-    z0: float
-    z_middle: float
-    t_in: float
-    t_out: float
-
-
-@dataclass
-class LithographyData(Part3DData):
-    """
-    :param thickness: The lithography thickness
-    :param z0: The starting z coordinate
-    :param layer_num: The layer number. Lower numbers go down first, with higher numbers
-        deposited last
-    :param litho_base: The base partNames to use. For multi-step lithography, the bases
-        are just all merged, so there is no need to list this more than once
-    """
-
-    thickness: float
-    z0: float
-    layer_num: int
-    litho_base: List[str]
-
-
-class Geom3DBuilder:
+class Geo3DBuilder:
     def extrusion(
         self,
         label: str,
         fc_name: Optional[str],
-        domain_type: str,
-        material: Optional[str],
         thickness: float,
         z0: float,
+        virtual: bool = False,
     ) -> ExtrudeData:
-        domain_type = Geom3DBuilder.convert_domain_type(domain_type)
         fc_name = label if not fc_name else fc_name
-        return ExtrudeData(domain_type, fc_name, label, material, thickness, z0)
+        return ExtrudeData(fc_name, label, virtual, thickness, z0)
 
     def wire(
         self,
         label: str,
         fc_name: Optional[str],
-        domain_type: str,
-        material: Optional[str],
         thickness: float,
         z0: float,
+        virtual: bool = False,
     ) -> WireData:
-        domain_type = Geom3DBuilder.convert_domain_type(domain_type)
         fc_name = label if not fc_name else fc_name
-        return WireData(domain_type, fc_name, label, material, thickness, z0)
+        return WireData(fc_name, label, virtual, thickness, z0)
 
     def wire_shell(
         self,
         label: str,
         fc_name: Optional[str],
-        domain_type: str,
-        material: Optional[str],
         thickness: float,
         target_wire: WireData,
         shell_verts: List[int],
         depo_mode: str,
+        virtual: bool = False,
     ) -> WireShellData:
-        domain_type = Geom3DBuilder.convert_domain_type(domain_type)
         fc_name = label if not fc_name else fc_name
         try:
             depo_mode = DepoMode[depo_mode]
@@ -187,68 +59,95 @@ class Geom3DBuilder:
                 f"{[d.name for d in DepoMode]}"
             )
         return WireShellData(
-            domain_type,
-            fc_name,
-            label,
-            material,
-            thickness,
-            target_wire,
-            shell_verts,
-            depo_mode,
+            fc_name, label, virtual, thickness, target_wire, shell_verts, depo_mode
         )
 
     def SAG(
         self,
         label: str,
         fc_name: Optional[str],
-        domain_type: str,
-        material: Optional[str],
         thickness: float,
         z0: float,
         z_middle: float,
         t_in: float,
         t_out: float,
+        virtual: bool = False,
     ) -> SAGData:
-        domain_type = Geom3DBuilder.convert_domain_type(domain_type)
         fc_name = label if not fc_name else fc_name
-        return SAGData(
-            domain_type, fc_name, label, material, thickness, z0, z_middle, t_in, t_out
-        )
+        return SAGData(fc_name, label, virtual, thickness, z0, z_middle, t_in, t_out)
 
     def lithography(
         self,
         label: str,
         fc_name: Optional[str],
-        domain_type: str,
-        material: Optional[str],
         thickness: float,
         z0: float,
         layer_num: int,
         litho_base: List[str],
+        virtual: bool = False,
     ) -> LithographyData:
-        domain_type = Geom3DBuilder.convert_domain_type(domain_type)
         fc_name = label if not fc_name else fc_name
         return LithographyData(
-            domain_type, fc_name, label, material, thickness, z0, layer_num, litho_base
+            fc_name, label, virtual, thickness, z0, layer_num, litho_base
         )
 
     def shape3d(
-        self,
-        label: str,
-        fc_name: Optional[str],
-        domain_type: str,
-        material: Optional[str],
+        self, label: str, fc_name: Optional[str], virtual: bool = False
     ) -> Part3DData:
-        domain_type = Geom3DBuilder.convert_domain_type(domain_type)
         fc_name = label if not fc_name else fc_name
-        return Part3DData(domain_type, fc_name, label, material)
+        return Part3DData(fc_name, label, virtual)
 
-    @staticmethod
-    def convert_domain_type(domain_str: str) -> DomainType:
-        try:
-            return DomainType[domain_str]
-        except KeyError:
+    def build_3d_geometry(
+        self,
+        input_parts,
+        input_file=None,
+        xsec_dict=None,
+        serialized_input_file=None,
+        params=None,
+    ):
+        """
+        Build a geometry in 3D.
+
+        :param list input_parts: Ordered list of input parts, leftmost items get built first
+        :param str input_file: Path to FreeCAD template file. Either this or serialized_input_file
+            must be set (but not both).
+        :param dict xsec_dict: Dictionary of cross-section specifications. It should be of the
+            form {'xsec_name':{'axis':(1,0,0),'distance':0.}}, where the axis parameter is a tuple
+            defining the axis that defines the normal of the cross section, and distance is
+            the length along the axis used to set the cross section.
+        :param bytes serialized_input_file: FreeCAD template file that has been serialized using
+            qmt.data.serialize_file. This is useful for passing a
+            file into a docker container or other environment that
+            doesn't have access to a shared drive. Either this or
+            serialized_input_file must be set (but not both).
+        :param dict params: Dictionary of parameters to use in FreeCAD.
+        :return Geo3DData: A built geometry.
+        """
+        if input_file is None and serialized_input_file is None:
             raise ValueError(
-                f"{domain_str} is not a valid domain type. Options are "
-                f"{[d.name for d in DomainType]}"
+                "One of input_file or serialized_input_file must be non-none."
             )
+        elif input_file is not None and serialized_input_file is not None:
+            raise ValueError("Both input_file and serialized_input_file were non-none.")
+        elif input_file is not None:
+            serial_fcdoc = serialize_file(input_file)
+        else:
+            serial_fcdoc = serialized_input_file
+        if params is None:
+            params = {}
+        if xsec_dict is None:
+            xsec_dict = {}
+        options_dict = {}
+        options_dict["serial_fcdoc"] = serial_fcdoc
+        options_dict["input_parts"] = input_parts
+        options_dict["params"] = params
+        options_dict["xsec_dict"] = xsec_dict
+
+        data = Geo3DData()
+        data.serial_fcdoc = serial_fcdoc
+        data.get_data("fcdoc")
+        from qmt.geometry.freecad.objectConstruction import build
+
+        built = build(options_dict)
+        FreeCAD.closeDocument("instance")
+        return built
