+++ qmt/materials/__init__.py
@@ -1,0 +1,1 @@
+from .materials import Materials
