+++ tests/test_builder_3d.py
@@ -1,7 +1,7 @@
-from qmt.geometry import Geom3DBuilder
+from qmt.geometry import Geo3DBuilder
 
 
 def test_builder_3d():
-    builder = Geom3DBuilder()
+    builder = Geo3DBuilder()
     extrude = builder.extrusion("blah", None, "virtual", None, 1, 2)
     pass
